AI_1_4/requirements.txt
requests>=2.28.0
pyyaml>=6.0
tiktoken>=0.5.0
